Move client logic to new package
