+++ lib/client.dart
@@ -1,1 +1,1 @@
-export "package:client/src/transport_client.dart";
+export "package:client/src/client/transport_client.dart";
