+++ lib/src/client/FileFactory.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:client/src/messages/file.dart';
+import 'package:client/src/client/messages/file.dart';
 
 class FileFactory {
   Future<File> createFile() async {
