+++ test/utils/stubs.dart
@@ -1,7 +1,7 @@
 import 'package:mockito/mockito.dart';
 
-import 'package:client/src/channel.dart';
-import 'package:client/src/messages/message.dart';
+import 'package:client/src/client/channel.dart';
+import 'package:client/src/client/messages/message.dart';
 
 class MockChannel extends Mock implements Channel {
   noSuchMethod(i) => super.noSuchMethod(i);
